requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-taggit==3.0.0
Pillow==9.2.0
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.4
